requirements.txt
openai>=1.12.0,<2.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdf2image==1.16.3
pillow>=10.3.0
